+++ templates/template-parts/social-link-layout.php
@@ -6,19 +6,20 @@
  */
 
  // Determine the page layout (i.e., grid or list).
- $layout = ( isset( $page_settings['type'] ) && '' !== $page_settings['type'] )
+ $layout  = ( isset( $page_settings['type'] ) && '' !== $page_settings['type'] )
  ? $page_settings['type']
  : 'grid';
+ $is_grid = 'grid' === $layout;
 ?>
 
 <!-- Set up list/grid of social links -->
-<h2 id="instagram-posts" class="hide-visually">
+<h2 id="instagram-posts" class="hide-visually" tabindex="-1">
  <?php esc_html_e( 'Instagram posts', 'gpalab-slo' ); ?>
 </h2>
 <div id="instructions">
- <p class="instructions <?php echo 'grid' !== $layout ? 'hide-visually' : ''; ?>">
+ <p class="instructions <?php echo ( ! $is_grid ) ? 'hide-visually' : ''; ?>">
     <?php
-      $item = 'grid' === $layout ? __( 'image', 'gpalab-slo' ) : __( 'item', 'gpalab-slo' );
+      $item = $is_grid ? __( 'image', 'gpalab-slo' ) : __( 'item', 'gpalab-slo' );
 
       /* translators: %s: list item type (one of image or item) */
       printf( esc_html__( 'Select an %s to see more', 'gpalab-slo' ), esc_html( $item ) );